allow configuration of motor arming condition
This allows the user to configure their own motor arming condition in
APM_Config.h. For example, the user could set it up for MODE1
transmitters, by using:

 #define MOTOR_ARM_CONDITION (g.rc_3.control_in == 0 && g.rc_1.control_in > 2700)
 #define MOTOR_DISARM_CONDITION (g.rc_3.control_in == 0 && g.rc_1.control_in < -2700)

You could also set it up to use the "throttle cut" switch on a
transmitter, by mapping that to a spare channel.

The default (throttle minimum and maximum/minimum yaw) is the same as
it was before

This patch also adds reporting of motor arming to the GCS, which can
be useful for testing.

git-svn-id: https://arducopter.googlecode.com/svn/trunk@1802 f9c3cf11-9bcb-44bc-f272-b75c42450872