Cleanup: rm ArduCopter/GCS_Mavlink\ copy.txt
* Not needed, but we can always checkout an old version if it is.

Hi Pat,
This may have been my WIP for the Mavlink 1.0 conversion. I tried to
give Tridge a head start on it. I've made small updates to it in
parallel to the current Mavlink. You can delete it as he would have used
it by now if he felt it was helpful.

Jason

On Feb 28, 2012, at 4:08 PM, Pat Hickey wrote:

It looks like this file was checked in by accident, possibly, but you've
made a few small changes to it since creating it.

Is it important? Should we get rid of it?

Git log of file below.

Thanks
Pat
...